safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A10_T4-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected token ';'. Expected ')' to end a compound expression.
S7.9_A10_T4

PASS Expected parsing failure

TEST COMPLETE
